+++ src/WillHero/GreenOrc.java
@@ -4,7 +4,7 @@ import javafx.scene.image.ImageView;
 
 public class GreenOrc extends Orc{
 
-    public GreenOrc(float x,float y,int health, int coin_Value, int push_Value, int jump_Value) {
+    public GreenOrc(double x,double y,int health, int coin_Value, int push_Value, int jump_Value) {
         super(x,y,health, coin_Value, push_Value, jump_Value);
     }
