+++ src/WillHero/Orc.java
@@ -6,7 +6,7 @@ public abstract class Orc extends GameComponents{
 	private int coin_Value;
 	private int push_Value;
 	private int jump_Value;
-	public Orc(float x,float y,int health,int coin_Value,int push_Value,int jump_Value) {
+	public Orc(double x,double y,int health,int coin_Value,int push_Value,int jump_Value) {
 		super(x,y);
 		this.health = health;
 		this.coin_Value = coin_Value;
