+++ src/WillHero/RedOrc.java
@@ -5,7 +5,7 @@ import javafx.scene.image.ImageView;
 public class RedOrc extends Orc {
 	private boolean isSleeping;
 	
-	public RedOrc(float x,float y,int health, int coin_Value, int push_Value, int jump_Value) {
+	public RedOrc(double x,double y,int health, int coin_Value, int push_Value, int jump_Value) {
 		super(x,y,health, coin_Value, push_Value, jump_Value);
 		this.setSleeping(true);
 	}
