+++ src/WillHero/WeaponChest.java
@@ -5,7 +5,7 @@ import javafx.scene.image.ImageView;
 public class WeaponChest extends GameComponents implements Chest{
     private Weapon rewardWeapon;
 
-    public WeaponChest(float x, float y,Weapon myWeapon) {
+    public WeaponChest(double x,double y,Weapon myWeapon) {
         super(x, y);
         this.rewardWeapon = myWeapon;
     }
